First version of simple batch inserter.
git-svn-id: https://svn.neo4j.org/components/neo/trunk@2753 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ enterprise/src/java/org/neo4j/impl/transaction/XidImpl.java
@@ -192,7 +192,7 @@ public class XidImpl implements Xid
         buf.append( "], BranchId[ " );
         for ( int i = 0; i < branchId.length; i++ )
         {
-            buf.append( (byte) branchId[i] + " " );
+            buf.append( branchId[i] + " " );
         }
         buf.append( "]" );
         return buf.toString();
